Normalize whitepace to use tab indentation
This is to comply with PostgreSQL convention for C code.
Tabulators (intended to be 4-space wide) are used for indentation.
The spacing has been adjusted so that tabs are used only for
indenting, while spaces are used to align with the previous line
beyond the indentation point (so that tab width can be changed
while preserving the code alignment).

See #5