Add the carto-package.json file
I added postgres 10.1.1 in reference to our fork. Although it shall be
possible to build against 9.5+ versions, some of them are known to
have subtle failures. See
https://github.com/CartoDB/odbc_fdw/issues/71 for an example

Mind that the intent of this file is the automation of the on-premise
dependencies and packaging. So instead of trying to cover all the
cases we shall focus on solving the problem at hand.